Référencer LVGL localement

+++ ui/ui.c
@@ -1,6 +1,7 @@
 #include "ui.h"
 #include <lvgl.h>
 #include "battery.h"
+#include <inttypes.h>
 
 static lv_style_t style_dark;
 
@@ -17,14 +18,14 @@ static void btn_event_cb(lv_event_t *e)
 static void slider_event_cb(lv_event_t *e)
 {
     lv_obj_t *slider = lv_event_get_target(e);
-    lv_label_set_text_fmt(label_slider, "%d", lv_slider_get_value(slider));
+    lv_label_set_text_fmt(label_slider, "%"PRId32"", (int32_t)lv_slider_get_value(slider));
 }
 
 static void battery_timer_cb(lv_timer_t *t)
 {
     battery_update();
     lv_bar_set_value(bar_batt, battery_get_percent(), LV_ANIM_OFF);
-    lv_label_set_text_fmt(label_batt, "%d%%", battery_get_percent());
+    lv_label_set_text_fmt(label_batt, "%"PRId32"%%", (int32_t)battery_get_percent());
 }
 
 void ui_init(void) {
@@ -49,7 +50,7 @@ void ui_init(void) {
     lv_obj_set_width(slider, w - 40);
     lv_obj_add_event_cb(slider, slider_event_cb, LV_EVENT_VALUE_CHANGED, NULL);
     label_slider = lv_label_create(cont);
-    lv_label_set_text_fmt(label_slider, "%d", lv_slider_get_value(slider));
+    lv_label_set_text_fmt(label_slider, "%"PRId32"", (int32_t)lv_slider_get_value(slider));
 
     /* Bouton simple */
     lv_obj_t *btn = lv_btn_create(cont);
@@ -63,7 +64,7 @@ void ui_init(void) {
     lv_bar_set_range(bar_batt, 0, 100);
     lv_bar_set_value(bar_batt, battery_get_percent(), LV_ANIM_OFF);
     label_batt = lv_label_create(cont);
-    lv_label_set_text_fmt(label_batt, "%d%%", battery_get_percent());
+    lv_label_set_text_fmt(label_batt, "%"PRId32"%%", (int32_t)battery_get_percent());
 
     /* Rafraîchissement périodique du niveau batterie */
     lv_timer_create(battery_timer_cb, 2000, NULL);
